refactor: initial refactor of card creation code into seperate function
