+++ index.html
@@ -23,6 +23,9 @@
           <div class="tool">
             <button class="button" id="save">Export</button>
           </div>
+          <div class="tool">
+            <button class="button" id="create-statement">Create Statement</button>
+          </div>
         </div>
         <ul class="notifications">
           <li>
